backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.3
